requirements.txt
fastapi>=0.110
uvicorn>=0.27
pydantic>=2.6
python-dotenv>=1.0
numpy>=1.24
scikit-learn>=1.3
rank-bm25>=0.2.2
openai>=1.0.0
faiss-cpu>=1.8.1
pydantic-settings